+++ malina-rollup.js
@@ -11,11 +11,11 @@ export default function malinaRollup(option = {}) {
         async transform(code, id) {
             if(!option.extension.some(ext => id.endsWith('.' + ext))) return null;
             let result;
-
-            let opts = Object.assign({
+            let opts = await loadConfig(id, {
+                ...option,
                 path: id,
                 name: id.match(/([^/\\]+)\.\w+$/)[1]
-            }, option);
+            });
             try {
                 let ctx = await malina.compile(code, opts);
                 result = ctx.result;
@@ -38,4 +38,27 @@ export default function malinaRollup(option = {}) {
             return content_cache[id] || null;
         }
     };
+}
+
+
+async function loadConfig(filename, option) {
+  const fs = await import('fs');
+  let localConfig, parts = filename.split(/[/\\]/);
+  for(let i = parts.length - 1; i > 1; i--) {
+    let local = parts.slice(0, i).join('/') + '/malina.config.js';
+    if(fs.existsSync(local)) {
+      localConfig = local;
+      break;
+    }
+  }
+
+  if(localConfig) {
+    const confFn = (await import(localConfig)).default;
+    if(typeof (confFn) == 'function') option = await confFn(option, filename);
+    else option = confFn;
+  }
+  if(!option.path) option.path = filename;
+  if(!option.name) option.name = option.path.match(/([^/\\]+)\.\w+$/)[1];
+
+  return option;
 }